+++ cookbooks/printers/CHANGELOG.md
@@ -5,6 +5,9 @@ This file is used to list changes made in each version of printerdrv.
 ## 0.5.0:
 
 * Updated with DAC100 specifics.
+* Updated HP Printer Drivers to 2.14
+* Updated Epson Printer Drivers to 2.14
+* Disabled printer sharing
 
 ## 0.4.0: